[Day 4] 15강 학습(유저 업데이트 API, 삭제 API 예외 처리하기)

+++ src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -43,14 +43,36 @@ public class UserController {
 		});
 	}
 
+	// TODO: 2024.05.01 존재하지 않는 유저를 업데이트, 삭제 요청할 때 어떻게 예외처리할까?
+	// TODO: 2024.05.01 jdbcTemplate query(), update()
+
+	/**
+	 * 1) 유저를 조회하고 조회에 성공하면 0을 반환
+	 * 2) 존재하지 않으면 에러 반환
+	 * @param request
+	 */
 	@PutMapping("/user")
 	public void updateUser(@RequestBody UserUpdateRequest request) {
+		String readSql = "SELECT * FROM user WHERE id = ?";
+		boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, request.getId()).isEmpty();
+
+		if (isUserNotExist) {
+			throw new IllegalArgumentException();
+		}
+
 		String sql = "UPDATE user SET name = ? WHERE id = ?";
 		jdbcTemplate.update(sql, request.getName(), request.getId());
 	}
 
 	@DeleteMapping("/user")
 	public void deleteUser(@RequestParam String name) {
+		String readSql = "SELECT * FROM user WHERE name = ?";
+		boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, name).isEmpty();
+
+		if (isUserNotExist) {
+			throw new IllegalArgumentException();
+		}
+
 		String sql = "DELETE FROM user WHERE name = ?";
 		jdbcTemplate.update(sql, name);
 	}
